Fixed core dump, middle button zoom
Last release of the year! Fixed a double-free memory problem that cause
a repeatable core dump. Added a new zooming mechanism for people without
wheel-mice. Click on a point with the middle button, move right with the
button down to zoom in and left to zoom out. Some further improvements
to stipples and line types. Better default process file. Thanks to Mark
Mathews for the zoom suggestion!