SET, ENV vars, PATH, shellout
The SET command is implemented. Piglet now inheirits environment variables
from its parent process. SET with no arguments will print all ENV variables.
SET with one argument will print the value of the named variable. SET with
two arguments will set the value of the named variable. Piglet now uses the
PATH variable when parsing commands. If the command is not an internal
Piglet command, then it will be searched for in the PATH definition. If a
UNIX executable is found in the PATH will be run with the standard output
sent to the Piglet text window. Any variables that you SET inside piglet
will be accessible inside the subshell with $VARNAME. Replaced getline()
with fgets() function to make the code run on Solaris.