Implemented autosave backup file "./cells/#<cellname>.d"
The autosave file is updated after every change to the current
cell.  It always is up-to-date.  Any save to the current cell
unlinks the autosave file.  If an autosave file exists when
a cell is re-edited in a new session, it means that either piglet
crashed or the edit was exited without saving.